# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PromptVault API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/profile:
    get:
      operationId: getUserProfile
      tags: [user]
      summary: Get user profile
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUserProfile
      tags: [user]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /user/saved-packs/{packId}:
    post:
      operationId: savePack
      tags: [user]
      summary: Save a pack to wishlist
      parameters:
        - name: packId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
    delete:
      operationId: unsavePack
      tags: [user]
      summary: Remove a pack from wishlist
      parameters:
        - name: packId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /user/downloads/{packId}:
    get:
      operationId: getDownloadUrl
      tags: [user]
      summary: Get download URL for a purchased pack
      parameters:
        - name: packId
          in: path
          required: true
          schema:
            type: integer
        - name: fileType
          in: query
          schema:
            type: string
            enum: [PDF, ZIP]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DownloadUrlResponse"
  /user/password:
    patch:
      operationId: updatePassword
      tags: [user]
      summary: Update password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePasswordBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Reviews
  /reviews/{packId}:
    get:
      operationId: getPackReviews
      tags: [reviews]
      summary: Get reviews for a pack
      parameters:
        - name: packId
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewListResponse"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review
      parameters:
        - name: packId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{packId}/{reviewId}:
    patch:
      operationId: updateReview
      tags: [reviews]
      summary: Update a review
      parameters:
        - name: packId
          in: path
          required: true
          schema:
            type: integer
        - name: reviewId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete a review
      parameters:
        - name: packId
          in: path
          required: true
          schema:
            type: integer
        - name: reviewId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Checkout
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # Auth
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        role:
          type: string
          enum: [BUYER, ADMIN, SUPER_ADMIN]
        status:
          type: string
          enum: [ACTIVE, SUSPENDED, PENDING, DELETED]
        emailVerifiedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - displayName
        - role
        - status
        - createdAt
    UpdateProfileBody:
      type: object
      properties:
        displayName:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string
    UpdatePasswordBody:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
      required:
        - currentPassword
        - newPassword

    # Category
...
    Review:
      type: object
      properties:
        id:
          type: integer
        packId:
          type: integer
        userId:
          type: integer
        userDisplayName:
          type: string
        userAvatarUrl:
          type: ["string", "null"]
        rating:
          type: integer
          minimum: 1
          maximum: 5
        title:
          type: ["string", "null"]
        body:
          type: ["string", "null"]
        isVerified:
          type: boolean
        helpfulCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - packId
        - userId
        - userDisplayName
        - rating
        - isVerified
        - helpfulCount
        - createdAt
    ReviewListResponse:
      type: object
      properties:
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
        avgRating:
          type: ["number", "null"]
        ratingDistribution:
          type: object
          additionalProperties:
            type: integer
      required:
        - reviews
        - total
        - page
        - limit
        - totalPages
    CreateReviewBody:
      type: object
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        title:
          type: string
        body:
          type: string
      required:
        - rating

    # Order
...
    DownloadUrlResponse:
      type: object
      properties:
        url:
          type: string
        fileType:
          type: string
        expiresAt:
          type: string
          format: date-time
      required:
        - url
        - fileType
        - expiresAt
...